# Improve heap iterators.