Mostly improve allocations but I'm planning to do more using those benchmark.

```
name                   old time/op    new time/op    delta
HeapIterator-16          3.87ms ± 6%    3.41ms ± 1%  -11.92%  (p=0.008 n=5+5)
HeapSampleIterator-16    2.09ms ± 1%    2.09ms ± 1%     ~     (p=0.421 n=5+5)

name                   old alloc/op   new alloc/op   delta
HeapIterator-16          10.5kB ± 0%     8.2kB ± 0%  -21.88%  (p=0.008 n=5+5)
HeapSampleIterator-16    8.40kB ± 0%    6.10kB ± 0%  -27.34%  (p=0.000 n=5+4)

name                   old allocs/op  new allocs/op  delta
HeapIterator-16            12.0 ± 0%       5.0 ± 0%  -58.33%  (p=0.008 n=5+5)
HeapSampleIterator-16      12.0 ± 0%       5.0 ± 0%  -58.33%  (p=0.008 n=5+5)
```

I want to introduce a difference between the need to dedupe vs order data.

The later doesn't requires to pop from the heap, we could actually use a sort.slice.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>